Put the size check in one module
