+++ src/ast/types.rs
@@ -457,15 +457,6 @@ impl Debug for TypeNode {
     }
 }
 
-/// The change that gives a type with no size a size, chosen by what makes its layout endless.
-enum NoSizeRemedy {
-    /// The fields lead back to the type: a pointer anywhere on the way round ends the descent.
-    Box,
-    /// The fields lead to the same type constructor at a larger argument: the descent ends once the
-    /// recursion stops enlarging the argument, whatever is boxed.
-    SameTypeArguments,
-}
-
 impl TypeNode {
     // Find the minimum node which includes the specified source code position.
     pub fn find_node_at(&self, pos: &SourcePos) -> Option<EndNode> {
@@ -1227,131 +1218,6 @@ impl TypeNode {
         field_types
             .iter()
             .all(|field_ty| field_ty.is_fully_unboxed(type_env))
-    }
-
-    /// Why a value of `self` has no size, given the types its layout came through, and `None` where
-    /// it has one.
-    ///
-    /// # Arguments
-    /// * `in_place` - the types `self` sits inside with no pointer in between, outermost first.
-    ///   Reaching one of them again is a value that contains itself.
-    /// * `across_pointers` - every type the layout came through, the ones behind a pointer included.
-    ///   Reaching a larger type of the same type constructor there has no end either: the same
-    ///   fields lead from that one to a larger one again.
-    pub(crate) fn no_size_cause(
-        self: &Arc<TypeNode>,
-        in_place: &[Arc<TypeNode>],
-        across_pointers: &[Arc<TypeNode>],
-    ) -> Option<String> {
-        if let Some(i) = in_place.iter().position(|ancestor| ancestor == self) {
-            let cause = format!("its unboxed fields reach `{}` itself", self.to_string());
-            return Some(self.format_no_size_error(&in_place[i..], cause, NoSizeRemedy::Box));
-        }
-        // A function value is a pair of pointers whatever it takes and returns, so its size is
-        // settled. Every function type shares the `->` constructor, so the growth of one function's
-        // argument would otherwise be read off another's.
-        if self.is_closure() || self.is_funptr() {
-            return None;
-        }
-        // The same type constructor with arguments that have grown: the fields that led from that
-        // one here lead on to a larger one again. A type merely appearing inside another (`Tree`
-        // inside `(Tree, Tree)`) is how an ordinary recursive type is written, and the walk ends
-        // there by meeting `Tree` a second time.
-        let my_app_seq = self.flatten_type_application();
-        let grows_from = |ancestor: &Arc<TypeNode>| {
-            if ancestor == self {
-                return false;
-            }
-            let their_app_seq = ancestor.flatten_type_application();
-            their_app_seq.len() == my_app_seq.len()
-                && their_app_seq[0] == my_app_seq[0]
-                && their_app_seq[1..]
-                    .iter()
-                    .zip(my_app_seq[1..].iter())
-                    .all(|(their_arg, my_arg)| their_arg.embeds_in(my_arg))
-        };
-        if let Some(i) = across_pointers.iter().position(grows_from) {
-            let cause = "its fields reach ever larger types".to_string();
-            return Some(self.format_no_size_error(
-                &across_pointers[i..],
-                cause,
-                NoSizeRemedy::SameTypeArguments,
-            ));
-        }
-        None
-    }
-
-    /// Whether this type is embedded in `other`: it appears there with its own shape intact, with
-    /// more type around it or inside its arguments. An argument grown this way is what tells a type
-    /// reached again at a larger argument from one reached at a smaller or unrelated one.
-    fn embeds_in(self: &Arc<TypeNode>, other: &Arc<TypeNode>) -> bool {
-        // Inside one of `other`'s parts.
-        let inside = match &other.ty {
-            Type::TyApp(fun, arg) => self.embeds_in(fun) || self.embeds_in(arg),
-            Type::AssocTy(_, args) => args.iter().any(|arg| self.embeds_in(arg)),
-            Type::TyVar(_) | Type::TyCon(_) => false,
-        };
-        if inside {
-            return true;
-        }
-        // The same shape at the top, each part embedded in the part facing it.
-        match (&self.ty, &other.ty) {
-            (Type::TyVar(my_var), Type::TyVar(their_var)) => my_var.name == their_var.name,
-            (Type::TyCon(my_tycon), Type::TyCon(their_tycon)) => my_tycon == their_tycon,
-            (Type::TyApp(my_fun, my_arg), Type::TyApp(their_fun, their_arg)) => {
-                my_fun.embeds_in(their_fun) && my_arg.embeds_in(their_arg)
-            }
-            (Type::AssocTy(my_assoc, my_args), Type::AssocTy(their_assoc, their_args)) => {
-                my_assoc == their_assoc
-                    && my_args.len() == their_args.len()
-                    && my_args
-                        .iter()
-                        .zip(their_args.iter())
-                        .all(|(my_arg, their_arg)| my_arg.embeds_in(their_arg))
-            }
-            _ => false,
-        }
-    }
-
-    /// The report for a type with no size: what its fields do, the way down to it from the type that
-    /// shows it, and the change that gives it a size.
-    fn format_no_size_error(
-        self: &Arc<TypeNode>,
-        ancestors: &[Arc<TypeNode>],
-        cause: String,
-        remedy: NoSizeRemedy,
-    ) -> String {
-        let descent = ancestors
-            .iter()
-            .chain([self])
-            .map(|ty| ty.to_string())
-            .collect::<Vec<_>>();
-        // A type holding itself directly is the whole story already, so the way down is spelled
-        // out only where it passes through another type.
-        let holds_itself = ancestors.iter().all(|ty| ty == self);
-        let way_down = if holds_itself {
-            String::new()
-        } else {
-            format!(
-                " ({})",
-                descent
-                    .iter()
-                    .map(|ty| format!("`{}`", ty))
-                    .collect::<Vec<_>>()
-                    .join(" -> ")
-            )
-        };
-        let remedy = match remedy {
-            NoSizeRemedy::Box if holds_itself => format!("Make `{}` boxed.", descent[0]),
-            NoSizeRemedy::Box => "Make one of these types boxed.".to_string(),
-            NoSizeRemedy::SameTypeArguments => {
-                "Give the recursive occurrence the same type arguments.".to_string()
-            }
-        };
-        format!(
-            "`{}` has no size: {}{}. {}",
-            descent[0], cause, way_down, remedy,
-        )
     }
 
     /// Whether a value of this type is one indivisible reference-counting unit — counted as a whole by
